requirements.txt
fastapi
uvicorn
python-pptx
python-multipart
aspose-slides
pyyaml
platformdirs